pyproject.toml
[project]
name = "optimize2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.0.0",
    "fastapi>=0.116.1",
    "openai>=1.100.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "pymongo>=4.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.11.7",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
]
